requirements.txt
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
cryptography==42.0.5
daphne==4.1.2
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
dnspython==2.6.1
psycopg==3.1.18
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
redis==5.0.3
requests==2.31.0
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.3
